Restructure into files and update readme with overview
