+++ namespaces.go
@@ -1,0 +1,36 @@
+package goxml
+
+// Namespaces - map of prefix to uri for namespaces
+var Namespaces = map[string]string{
+	"algsupport": "urn:oasis:names:tc:SAML:metadata:algsupport",
+	"corto":      "http://corto.wayf.dk",
+	"ds":         "http://www.w3.org/2000/09/xmldsig#",
+	"idpdisc":    "urn:oasis:names:tc:SAML:profiles:SSO:idp-discovery-protocol",
+	"init":       "urn:oasis:names:tc:SAML:profiles:SSO:request-init",
+	"md":         "urn:oasis:names:tc:SAML:2.0:metadata",
+	"mdattr":     "urn:oasis:names:tc:SAML:metadata:attribute",
+	"mdrpi":      "urn:oasis:names:tc:SAML:metadata:rpi",
+	"mdui":       "urn:oasis:names:tc:SAML:metadata:ui",
+	"saml":       "urn:oasis:names:tc:SAML:2.0:assertion",
+	"saml1":      "urn:oasis:names:tc:SAML:1.0:assertion",
+	"samlp":      "urn:oasis:names:tc:SAML:2.0:protocol",
+	"sdss":       "http://sdss.ac.uk/2006/06/WAYF",
+	"shibmd":     "urn:mace:shibboleth:metadata:1.0",
+	"SOAP-ENV":   "http://schemas.xmlsoap.org/soap/envelope/",
+	"ukfedlabel": "http://ukfederation.org.uk/2006/11/label",
+	"wayf":       "http://wayf.dk/2014/08/wayf",
+	"xenc":       "http://www.w3.org/2001/04/xmlenc#",
+	"xenc11":     "http://www.w3.org/2009/xmlenc11#",
+	"xml":        "http://www.w3.org/XML/1998/namespace",
+	"xs":         "http://www.w3.org/2001/XMLSchema",
+	"xsi":        "http://www.w3.org/2001/XMLSchema-instance",
+	"xsl":        "http://www.w3.org/1999/XSL/Transform",
+	"ec":         "http://www.w3.org/2001/10/xml-exc-c14n#",
+	"aslo":       "urn:oasis:names:tc:SAML:2.0:protocol:ext:async-slo",
+	"t":          "http://schemas.xmlsoap.org/ws/2005/02/trust",
+	"wsu":        "http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd",
+	"wsp":        "http://schemas.xmlsoap.org/ws/2004/09/policy",
+	"wsa":        "http://www.w3.org/2005/08/addressing",
+	"a":          "http://schemas.xmlsoap.org/ws/2009/09/identity/claims",
+	"eidas":      "http://eidas.europa.eu/saml-extensions",
+}
